# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ImageFlow public OpenAPI protected by AK/SK HMAC authentication.",
        "title": "ImageFlow OpenAPI",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/openapi/cleanup": {
            "post": {
                "description": "Trigger expired image cleanup. Requires api:cleanup permission.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OpenAPI"
                ],
                "summary": "Trigger cleanup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access Key",
                        "name": "X-Access-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "HMAC-SHA256 signature",
                        "name": "X-Signature",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unix timestamp",
                        "name": "X-Timestamp",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/openapi/config": {
            "get": {
                "description": "Get public client configuration. Requires api:config permission.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OpenAPI"
                ],
                "summary": "Get client configuration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access Key",
                        "name": "X-Access-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "HMAC-SHA256 signature",
                        "name": "X-Signature",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unix timestamp",
                        "name": "X-Timestamp",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.ClientConfig"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/openapi/debug/tags": {
            "get": {
                "description": "Debug tag indexes. Requires api:debug permission.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OpenAPI"
                ],
                "summary": "Debug tags",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access Key",
                        "name": "X-Access-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "HMAC-SHA256 signature",
                        "name": "X-Signature",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unix timestamp",
                        "name": "X-Timestamp",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/openapi/delete": {
            "post": {
                "description": "Delete an image. Requires api:delete permission.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OpenAPI"
                ],
                "summary": "Delete image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access Key",
                        "name": "X-Access-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "HMAC-SHA256 signature",
                        "name": "X-Signature",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unix timestamp",
                        "name": "X-Timestamp",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Delete request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/openapi/images": {
            "get": {
                "description": "List uploaded images. Requires api:images permission.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OpenAPI"
                ],
                "summary": "List images",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access Key",
                        "name": "X-Access-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "HMAC-SHA256 signature",
                        "name": "X-Signature",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unix timestamp",
                        "name": "X-Timestamp",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by tag",
                        "name": "tag",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/openapi/random": {
            "get": {
                "description": "Get a random image with optional filters. Requires api:random permission.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OpenAPI"
                ],
                "summary": "Get random image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access Key",
                        "name": "X-Access-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "HMAC-SHA256 signature",
                        "name": "X-Signature",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unix timestamp",
                        "name": "X-Timestamp",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Single tag filter",
                        "name": "tag",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated tag filters",
                        "name": "tags",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated excluded tags",
                        "name": "exclude",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "portrait or landscape",
                        "name": "orientation",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "avif, webp, or original",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/openapi/tags": {
            "get": {
                "description": "List available image tags. Requires api:tags permission.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OpenAPI"
                ],
                "summary": "List tags",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access Key",
                        "name": "X-Access-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "HMAC-SHA256 signature",
                        "name": "X-Signature",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unix timestamp",
                        "name": "X-Timestamp",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/openapi/upload": {
            "post": {
                "description": "Upload images through OpenAPI. Requires api:upload permission.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OpenAPI"
                ],
                "summary": "Upload images",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access Key",
                        "name": "X-Access-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "HMAC-SHA256 signature",
                        "name": "X-Signature",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unix timestamp",
                        "name": "X-Timestamp",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Image files",
                        "name": "images",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated tags",
                        "name": "tags",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Expiration in minutes",
                        "name": "expiryMinutes",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.UploadResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "config.ClientConfig": {
            "type": "object",
            "properties": {
                "avifSupport": {
                    "description": "Whether AVIF format is supported",
                    "type": "boolean"
                },
                "imageQuality": {
                    "description": "Image conversion quality (1-100)",
                    "type": "integer"
                },
                "maxUploadCount": {
                    "description": "Maximum number of images allowed per upload",
                    "type": "integer"
                },
                "speed": {
                    "description": "Encoding speed (0-8, 0=slowest/highest quality)",
                    "type": "integer"
                }
            }
        },
        "errors.ErrorCode": {
            "type": "integer",
            "enum": [
                1000,
                1001,
                1002,
                1003,
                1004,
                2000,
                2001,
                2002,
                2003,
                2004,
                3001,
                3002,
                3003
            ],
            "x-enum-comments": {
                "ErrAKSKDisabled": "AK/SK is disabled",
                "ErrForbidden": "Forbidden",
                "ErrImageDelete": "Image deletion error",
                "ErrImageList": "Image list retrieval error",
                "ErrImageProcess": "Image processing error",
                "ErrImageUpload": "Image upload error",
                "ErrInternal": "Internal server error",
                "ErrInvalidAKSK": "Invalid AK/SK credentials",
                "ErrInvalidParam": "Invalid parameter",
                "ErrMetadata": "Metadata operation error",
                "ErrNotFound": "Resource not found",
                "ErrPermDenied": "Permission denied for AK/SK",
                "ErrUnauthorized": "Unauthorized"
            },
            "x-enum-descriptions": [
                "Internal server error",
                "Invalid parameter",
                "Unauthorized",
                "Forbidden",
                "Resource not found",
                "Image processing error",
                "Image upload error",
                "Image deletion error",
                "Image list retrieval error",
                "Metadata operation error",
                "Invalid AK/SK credentials",
                "AK/SK is disabled",
                "Permission denied for AK/SK"
            ],
            "x-enum-varnames": [
                "ErrInternal",
                "ErrInvalidParam",
                "ErrUnauthorized",
                "ErrForbidden",
                "ErrNotFound",
                "ErrImageProcess",
                "ErrImageUpload",
                "ErrImageDelete",
                "ErrImageList",
                "ErrMetadata",
                "ErrInvalidAKSK",
                "ErrAKSKDisabled",
                "ErrPermDenied"
            ]
        },
        "errors.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Error code",
                    "allOf": [
                        {
                            "$ref": "#/definitions/errors.ErrorCode"
                        }
                    ]
                },
                "details": {
                    "description": "Error details"
                },
                "message": {
                    "description": "Error message",
                    "type": "string"
                }
            }
        },
        "handlers.UploadResponse": {
            "type": "object",
            "properties": {
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.UploadResult"
                    }
                }
            }
        },
        "handlers.UploadResult": {
            "type": "object",
            "properties": {
                "expiryTime": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "format": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "orientation": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "urls": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "AkSkAccessKey": {
            "type": "apiKey",
            "name": "X-Access-Key",
            "in": "header"
        },
        "AkSkSignature": {
            "type": "apiKey",
            "name": "X-Signature",
            "in": "header"
        },
        "AkSkTimestamp": {
            "type": "apiKey",
            "name": "X-Timestamp",
            "in": "header"
        }
    }
}